# Make resolution thread-safe