--- a/BoDi.Performance.Tests/BoDi.Performance.Tests.csproj
+++ b/BoDi.Performance.Tests/BoDi.Performance.Tests.csproj
@@ -1,0 +1,16 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFrameworks>netcoreapp3.1</TargetFrameworks>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="BenchmarkDotNet" Version="0.12.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\BoDi\BoDi.csproj" />
+  </ItemGroup>
+
+</Project>