--- a/BoDi.Tests/RegisterFactoryDelegateTests.cs
+++ b/BoDi.Tests/RegisterFactoryDelegateTests.cs
@@ -86,7 +86,7 @@ namespace BoDi.Tests
 
         [Test/*, ExpectedException(typeof(ObjectContainerException), ExpectedMessage = "Circular dependency", MatchType = MessageMatch.Contains)*/]
         [Ignore("dynamic circles not detected yet, this leads to stack overflow")]
-        public void ShouldThrowExceptionForDynamicCircuarDepenencies()
+        public void ShouldThrowExceptionForDynamicCircularDependencies()
         {
             // given
 
@@ -98,7 +98,7 @@ namespace BoDi.Tests
         }
 
         [Test/*, ExpectedException(typeof(ObjectContainerException), ExpectedMessage = "Circular dependency", MatchType = MessageMatch.Contains)*/]
-        public void ShouldThrowExceptionForStaticCircuarDepenencies()
+        public void ShouldThrowExceptionForStaticCircularDependencies()
         {
             // given
 
@@ -111,7 +111,7 @@ namespace BoDi.Tests
         }
 
         [Test/*, ExpectedException(typeof(ObjectContainerException), ExpectedMessage = "Circular dependency", MatchType = MessageMatch.Contains)*/]
-        public void ShouldThrowExceptionForStaticCircuarDepenenciesWithMultipleFactoriesInPath()
+        public void ShouldThrowExceptionForStaticCircularDependenciesWithMultipleFactoriesInPath()
         {
             // given