… BoDi.Performance.Tests/1.4/BoDi.cs
… BoDi.Performance.Tests/1.next-flawed/BoDi.cs
… BoDi.Performance.Tests/Benchmarks/BenchmarkBase.cs
… BoDi.Performance.Tests/Benchmarks/MultipleContainerBenchmarkBase.cs
… BoDi.Performance.Tests/Benchmarks/ResolveAllFromFactory.cs
… BoDi.Performance.Tests/Benchmarks/ResolveAllFromFactoryRepeated.cs
… BoDi.Performance.Tests/Benchmarks/ResolveAllFromType.cs
… BoDi.Performance.Tests/Benchmarks/ResolveAllFromTypeRepeated.cs
… BoDi.Performance.Tests/Benchmarks/ResolveFromFactory.cs
… BoDi.Performance.Tests/Benchmarks/ResolveFromFactoryRepeated.cs
… BoDi.Performance.Tests/Benchmarks/ResolveFromType.cs
… BoDi.Performance.Tests/Benchmarks/ResolveFromTypeRepeated.cs
… BoDi.Performance.Tests/Benchmarks/ResolveMultiThreadedMultipleContainer.cs
… BoDi.Performance.Tests/Benchmarks/ResolveMultiThreadedSingleContainer.cs
… BoDi.Performance.Tests/Benchmarks/SingleContainerBenchmarkBase.cs
… BoDi.Tests/ResolveTests.cs
… BoDi.sln
… BoDi/BoDi.cs