Hi @gasparnagy ,

Following the discussions on BoDi thread-safety, I just took the time to write
* unit-tests demonstrating the issue
* code fixing these unit tests for resolution on the container level, ensuring locks are taken once, on the publicly exposed API, not impacting other internal calls, or not taking the several locks when using a container hierarchy. Note that:
    * #24 was aiming for registration and resolution thread-safety, but I assume it is safe to manage resolution only
    * #31 was only targeting base container access
* performance benchmark code to quantify performance impact


You will find below the results of the benchmarks.
As you can see, impact ranges from 3 to 7%, but this seems a very acceptable trade-off to me to gain thread safety, especially as resolution times are actually probably quite low with regards to the run time of code relying on BoDi, like SpecFlow.

Would be happy to have your view on this proposal!

``` ini
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
Intel Core i7-4790 CPU 3.60GHz (Haswell), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=5.0.100
  [Host]     : .NET Core 3.1.10 (CoreCLR 4.700.20.51601, CoreFX 4.700.20.51901), X64 RyuJIT
  DefaultJob : .NET Core 3.1.10 (CoreCLR 4.700.20.51601, CoreFX 4.700.20.51901), X64 RyuJIT
```
**ResolveFromType**
|  Method |     Mean |   Error |  StdDev |      Min |      Max |   Median | Ratio | Rank |
|-------- |---------:|--------:|--------:|---------:|---------:|---------:|------:|-----:|
|    v1.4 | 209.3 ns | 0.56 ns | 0.50 ns | 208.3 ns | 210.0 ns | 209.3 ns |  1.00 |    1 |
| Current | 223.0 ns | 0.55 ns | 0.51 ns | 221.9 ns | 223.8 ns | 223.1 ns |  1.07 |    2 |

**ResolveFromFactory**
|  Method |     Mean |   Error |  StdDev |      Min |      Max |   Median | Ratio | Rank |
|-------- |---------:|--------:|--------:|---------:|---------:|---------:|------:|-----:|
|    v1.4 | 226.2 ns | 1.38 ns | 1.15 ns | 224.5 ns | 228.4 ns | 226.1 ns |  1.00 |    1 |
| Current | 233.9 ns | 0.75 ns | 0.70 ns | 233.0 ns | 235.1 ns | 234.0 ns |  1.03 |    2 |

**ResolveAllFromType**
|  Method |     Mean |   Error |  StdDev |      Min |      Max |   Median | Ratio | Rank |
|-------- |---------:|--------:|--------:|---------:|---------:|---------:|------:|-----:|
|    v1.4 | 486.7 ns | 1.75 ns | 1.55 ns | 484.3 ns | 488.5 ns | 487.2 ns |  1.00 |    1 |
| Current | 516.8 ns | 1.55 ns | 1.21 ns | 513.4 ns | 518.0 ns | 517.0 ns |  1.06 |    2 |

**ResolveAllFromFactory**
|  Method |     Mean |   Error |  StdDev |      Min |      Max |   Median | Ratio | Rank |
|-------- |---------:|--------:|--------:|---------:|---------:|---------:|------:|-----:|
|    v1.4 | 499.4 ns | 1.72 ns | 1.53 ns | 496.5 ns | 502.3 ns | 499.4 ns |  1.00 |    1 |
| Current | 512.8 ns | 2.30 ns | 2.16 ns | 510.7 ns | 517.5 ns | 512.0 ns |  1.03 |    2 |