@Greybird 2020-12-13
Hello @gasparnagy,

Would you have time to review this proposal and provide feedback ?

Thanks in advance,