@tzongithub 2021-01-22
Hello @Greybird,

thank you for your contribution!
We've discussed with @gasparnagy that - BoDi being a fundamental base component of SpecFlow - the SpecFlow team will look into these issues now. In our current OSS Sprint we've selected 3 parallel execution issues from our backlog: https://github.com/SpecFlowOSS/SpecFlow/issues/2224. I think that some of these issues might be caused by the fact that the container is not thread-safe, however, they are not easily reproducible. So I'm looking into potential solutions including the question of BoDi thread-safety. ;)

I've looked into your proposed solution with @alexandrafung and here are our thoughts:

* We really like the unit tests: you found a nice way to force the synchronization of the ctors. Hence your concurrency tests can reproduce the object creation race condition in a deterministic way. Great job!
* The performance benchmark code is also nice, but here we have some concerns if we are measuring the "real life" performance penalty of the synchronization. The benchmark measures the overhead of the "lock" calls if running on a single thread. However, a more significant penalty might occur when several threads try to resolve an object from the container and have to queue up for access. 

With the solution of locking the resolve operations on the public interface level we see 2 problems:
1. This approach locks "resolve" access to the container and forces all threads to queue up for resolving. This is true not only when the object needs to be created and saved in the container, but also when the objects is already created and should be simply returned. I think this causes much overhead during execution. To phrase it another way: not only the "write" access is synchronized, but also the (otherwise thread-safe) "read" access, which causes unnecessary overhead.
2. The current approach locks the resolve access on container level. However, SpecFlow works with a hierarchy of containers (scenario-feature-test thread-global). If you resolve a dependency from the scenario container that is registered in the global container, the scenario container will lock "resolve" access on itself, but not on the global container, where the registration is found. Hence, if another parallel thread resolves the same dependency from its own scenario container it will not be blocked (as the lock is on the other scenario container instance) and the global dependency will be resolved from the global container in a race condition. 

Here is a test case that demonstrates the second problem (for the sake of simplicity the other resolve call is done on the base container, and not on another child container):

``` csharp
        [Test]
        public void ShouldBeAbleToResolveFromMultipleThreadsFromContainerHierarchyWhenRegisteredAsType(
            [Values(RegistrationStrategy.PerContext, RegistrationStrategy.PerDependency)] RegistrationStrategy registrationStrategy,
            [Values(null, "Name")] string name)
        {
            try
            {
                
                IObjectContainer baseContainer = new ObjectContainer();
                IObjectContainer childContainer = new ObjectContainer(baseContainer);
                var registration = baseContainer.RegisterTypeAs<BlockingObject, BlockingObject>(name);
                
                ApplyRegistrationStrategy(registration, registrationStrategy);

                var thread1 = new Thread(_ => Assert.That(() => baseContainer.Resolve<BlockingObject>(name), Throws.Nothing));
                var thread2 = new Thread(_ => Assert.That(() => childContainer.Resolve<BlockingObject>(name), Throws.Nothing));
                thread1.Start();
                thread2.Start();

                // try to wait until both object constructions are in progress (may not happen if no threading issue)
                BlockingObject.ObjectsCreated.WaitOne(ForHalfASecond);
                BlockingObject.ObjectsCreated.WaitOne(ForHalfASecond);
                
                // allow constructors to finish
                BlockingObject.ConstructorBlockers[0].Set();
                BlockingObject.ConstructorBlockers[1].Set();

                // complete the threads
                if (!thread1.Join(ForHalfASecond))
                {
                    Assert.Fail("Unable to complete resolution");
                }
                if (!thread2.Join(ForHalfASecond))
                {
                    Assert.Fail("Unable to complete resolution");
                }
            }
            finally
            {
                BlockingObject.ResetBlockingEvents();
            }
        }
```

We're considering and alternative solution, where the lock would be applied only when a registration attempts to create a new object. I think this could solve the above mentioned issues, however, it is a deeper change and brings some interesting edge cases with it.

What do you think?