@Greybird 2021-01-23
Hi @tzongithub,

Thanks for the interest in my proposal, and for the very detailed feedback!
I should have thought to the container hierarchy when I decided to reduce the scope of the lock :sweat:

For the performance tests, the rationale was that currently, multi-threading was not supported, and as such, not comparable, and the key was to check the performance degradation on the single thread scenario caused by the code introduced to make the container thread-safe.
However, I fully agree with you that the lock will introduce queuing in a way.

I've tried to enhance my proposal with more tests (both unit - including the one you provided) and performance ones, for multiple chained resolution, and multi-threaded resolution, on both single container, and on a container hierarchy.
I also tried to implement a read-only lock, by combining `ConcurrentDictionary`, which is lock-free on read, with a `Lazy`, to ensure the delegate used by its `GetOrAdd` method is only creating a single object.
It seems impact on first resolution is higher, but for subsequent resolutions, it seems to behave quite ok, from what I can tell.

Now, I see (at least) the remaining issue:
It is possible to change the registrations during the lifetime of a container, and this code is not thread-safe.
* the `AssertNotResolved` method/`resolvedKeys` can fail to report an issue in multi thread scenarios
* the fact that we have a multi step resolution and registration (with the `NamedInstanceDictionaryRegistration`) is making the issue hard to solve
Maybe switching to a pattern similar to Microsoft dependency injection (`IServiceCollection` => `IServiceProvider`) would help, but it would also be a breaking change.

Do you have a view on this?

Here are the benchmark results (only version that can support the scenarios without failing are used):

``` ini
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
Intel Core i7-4790 CPU 3.60GHz (Haswell), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=5.0.102
  [Host]     : .NET Core 3.1.11 (CoreCLR 4.700.20.56602, CoreFX 4.700.20.56604), X64 RyuJIT
  DefaultJob : .NET Core 3.1.11 (CoreCLR 4.700.20.56602, CoreFX 4.700.20.56604), X64 RyuJIT
```
**ResolveFromType**
|         Method |     Mean |   Error |  StdDev |      Min |      Max |   Median | Ratio | RatioSD | Rank |
|--------------- |---------:|--------:|--------:|---------:|---------:|---------:|------:|--------:|-----:|
|           v1.4 | 219.9 ns | 2.76 ns | 2.45 ns | 217.3 ns | 225.2 ns | 219.3 ns |  1.00 |    0.00 |    1 |
| v1.Next-Flawed | 229.3 ns | 0.87 ns | 0.77 ns | 227.7 ns | 230.6 ns | 229.3 ns |  1.04 |    0.01 |    2 |
|        Current | 361.5 ns | 1.71 ns | 1.43 ns | 358.6 ns | 363.0 ns | 361.6 ns |  1.64 |    0.02 |    3 |


**ResolveFromFactory**
|         Method |     Mean |   Error |  StdDev |      Min |      Max |   Median | Ratio | Rank |
|--------------- |---------:|--------:|--------:|---------:|---------:|---------:|------:|-----:|
|           v1.4 | 233.8 ns | 1.26 ns | 1.18 ns | 231.7 ns | 235.6 ns | 234.0 ns |  1.00 |    1 |
| v1.Next-Flawed | 243.6 ns | 1.38 ns | 1.16 ns | 241.8 ns | 246.4 ns | 243.7 ns |  1.04 |    2 |
|        Current | 355.0 ns | 2.04 ns | 1.81 ns | 352.3 ns | 358.3 ns | 354.6 ns |  1.52 |    3 |

**ResolveAllFromType**
|         Method |      Mean |    Error |   StdDev |       Min |       Max |    Median | Ratio | Rank |
|--------------- |----------:|---------:|---------:|----------:|----------:|----------:|------:|-----:|
|        Current |  73.33 ns | 0.960 ns | 0.851 ns |  72.44 ns |  75.23 ns |  73.09 ns |  0.15 |    1 |
|           v1.4 | 503.20 ns | 3.558 ns | 3.328 ns | 497.76 ns | 510.85 ns | 502.72 ns |  1.00 |    2 |
| v1.Next-Flawed | 537.39 ns | 5.017 ns | 4.189 ns | 533.53 ns | 548.47 ns | 536.47 ns |  1.07 |    3 |

**ResolveAllFromFactory**
|         Method |      Mean |    Error |   StdDev |       Min |       Max |    Median | Ratio | RatioSD | Rank |
|--------------- |----------:|---------:|---------:|----------:|----------:|----------:|------:|--------:|-----:|
|        Current |  73.11 ns | 0.554 ns | 0.491 ns |  72.40 ns |  74.07 ns |  73.08 ns |  0.14 |    0.00 |    1 |
|           v1.4 | 514.87 ns | 4.131 ns | 3.662 ns | 511.24 ns | 522.61 ns | 513.58 ns |  1.00 |    0.00 |    2 |
| v1.Next-Flawed | 528.64 ns | 7.337 ns | 6.126 ns | 522.54 ns | 542.15 ns | 526.95 ns |  1.03 |    0.02 |    3 |

**ResolveFromTypeRepeated**
|         Method | Repetitions |     Mean |     Error |    StdDev |      Min |      Max |   Median | Ratio | Rank |
|--------------- |------------ |---------:|----------:|----------:|---------:|---------:|---------:|------:|-----:|
|        Current |          10 | 3.659 μs | 0.0208 μs | 0.0184 μs | 3.625 μs | 3.690 μs | 3.660 μs |  0.98 |    1 |
| v1.Next-Flawed |          10 | 3.673 μs | 0.0187 μs | 0.0146 μs | 3.658 μs | 3.706 μs | 3.669 μs |  0.98 |    1 |
|           v1.4 |          10 | 3.742 μs | 0.0239 μs | 0.0224 μs | 3.709 μs | 3.784 μs | 3.743 μs |  1.00 |    2 |

**ResolveFromFactoryRepeated**
|         Method | Repetitions |     Mean |     Error |    StdDev |      Min |      Max |   Median | Ratio | Rank |
|--------------- |------------ |---------:|----------:|----------:|---------:|---------:|---------:|------:|-----:|
|           v1.4 |          10 | 2.322 μs | 0.0093 μs | 0.0078 μs | 2.308 μs | 2.333 μs | 2.324 μs |  1.00 |    1 |
| v1.Next-Flawed |          10 | 2.429 μs | 0.0230 μs | 0.0216 μs | 2.400 μs | 2.475 μs | 2.421 μs |  1.05 |    2 |
|        Current |          10 | 3.580 μs | 0.0162 μs | 0.0152 μs | 3.558 μs | 3.610 μs | 3.578 μs |  1.54 |    3 |

**ResolveAllFromTypeRepeated**
|         Method | Repetitions |       Mean |    Error |   StdDev |        Min |        Max |     Median | Ratio | Rank |
|--------------- |------------ |-----------:|---------:|---------:|-----------:|-----------:|-----------:|------:|-----:|
|        Current |          10 |   728.7 ns |  3.35 ns |  2.97 ns |   723.6 ns |   735.1 ns |   728.2 ns |  0.14 |    1 |
|           v1.4 |          10 | 5,048.0 ns | 37.08 ns | 32.87 ns | 5,012.1 ns | 5,116.7 ns | 5,039.0 ns |  1.00 |    2 |
| v1.Next-Flawed |          10 | 5,154.3 ns | 40.91 ns | 34.16 ns | 5,115.0 ns | 5,243.7 ns | 5,147.3 ns |  1.02 |    3 |

**ResolveAllFromFactoryRepeated**
|         Method | Repetitions |       Mean |    Error |   StdDev |        Min |        Max |     Median | Ratio | Rank |
|--------------- |------------ |-----------:|---------:|---------:|-----------:|-----------:|-----------:|------:|-----:|
|        Current |          10 |   721.5 ns |  5.22 ns |  4.88 ns |   715.9 ns |   730.5 ns |   719.6 ns |  0.14 |    1 |
|           v1.4 |          10 | 5,117.6 ns | 35.89 ns | 31.82 ns | 5,087.0 ns | 5,193.3 ns | 5,107.4 ns |  1.00 |    2 |
| v1.Next-Flawed |          10 | 5,320.3 ns | 49.13 ns | 43.55 ns | 5,269.1 ns | 5,426.3 ns | 5,322.5 ns |  1.04 |    3 |

**ResolveMultiThreadedSingleContainer**
|         Method | ThreadCount | Loops |     Mean |     Error |    StdDev |   Median |      Min |      Max | Rank |
|--------------- |------------ |------ |---------:|----------:|----------:|---------:|---------:|---------:|-----:|
|        Current |           4 |  2048 | 2.338 ms | 0.0456 ms | 0.0639 ms | 2.335 ms | 2.205 ms | 2.476 ms |    1 |
|        Current |           8 |  2048 | 2.344 ms | 0.0448 ms | 0.0516 ms | 2.341 ms | 2.261 ms | 2.445 ms |    1 |
|        Current |          16 |  2048 | 2.802 ms | 0.0559 ms | 0.1586 ms | 2.713 ms | 2.624 ms | 3.378 ms |    2 |
|        Current |           2 |  2048 | 3.306 ms | 0.0255 ms | 0.0213 ms | 3.308 ms | 3.271 ms | 3.354 ms |    3 |
| v1.Next-Flawed |           2 |  2048 | 5.079 ms | 0.0225 ms | 0.0211 ms | 5.084 ms | 5.030 ms | 5.108 ms |    4 |
| v1.Next-Flawed |           4 |  2048 | 5.284 ms | 0.0155 ms | 0.0137 ms | 5.283 ms | 5.259 ms | 5.309 ms |    5 |
| v1.Next-Flawed |          16 |  2048 | 5.811 ms | 0.0136 ms | 0.0120 ms | 5.814 ms | 5.788 ms | 5.828 ms |    6 |
| v1.Next-Flawed |           8 |  2048 | 5.881 ms | 0.0193 ms | 0.0171 ms | 5.878 ms | 5.850 ms | 5.916 ms |    6 |

**ResolveMultiThreadedMultipleContainer**
|  Method | ThreadCount | Loops |     Mean |     Error |    StdDev |      Min |      Max |   Median | Rank |
|-------- |------------ |------ |---------:|----------:|----------:|---------:|---------:|---------:|-----:|
| Current |           4 |  2048 | 3.130 ms | 0.0592 ms | 0.0790 ms | 2.957 ms | 3.275 ms | 3.147 ms |    1 |
| Current |           8 |  2048 | 3.150 ms | 0.0622 ms | 0.0740 ms | 3.037 ms | 3.268 ms | 3.158 ms |    1 |
| Current |          16 |  2048 | 3.524 ms | 0.0509 ms | 0.0476 ms | 3.467 ms | 3.596 ms | 3.516 ms |    2 |
| Current |           2 |  2048 | 4.834 ms | 0.0420 ms | 0.0393 ms | 4.761 ms | 4.897 ms | 4.840 ms |    3 |